Mostly Figure 2
Look over Fig1 Code to see if any changes need to be made.
Look over Fig2 Code fixes to make it run smoothly. The CSV's changed
formatt slightly. I also had mispelled "control" when reading the data.
The loop length statement needed to change to the "filter". and idicator
for filer was set to the third column.  Changed labels for fig 2. Start
debug on Fig 3